// src/data/groups.generated.ts
// AUTO-GENERATED FILE. DO NOT EDIT.
// Generated from data/affiliations.csv by scripts/generate-groups.cjs

export interface GroupRecord {
  groupId: string;
  shortName: string;
  name: string;
  affiliationShort: string;
  affiliationName: string;
}

export const groups: GroupRecord[] = [
  {
    "groupId": "OFC",
    "shortName": "OFC",
    "name": "Oilfield Chemistry",
    "affiliationShort": "CIPR",
    "affiliationName": "Center for Integrative Petroleum Research"
  },
  {
    "groupId": "MMM",
    "shortName": "MMM",
    "name": "Multiscale and Multiphysics Modeling",
    "affiliationShort": "CIPR",
    "affiliationName": "Center for Integrative Petroleum Research"
  },
  {
    "groupId": "PRP",
    "shortName": "PRP",
    "name": "Petrophysics & Rock Physics",
    "affiliationShort": "CIPR",
    "affiliationName": "Center for Integrative Petroleum Research"
  },
  {
    "groupId": "GP",
    "shortName": "GP",
    "name": "Geophysics",
    "affiliationShort": "CIPR",
    "affiliationName": "Center for Integrative Petroleum Research"
  },
  {
    "groupId": "GS",
    "shortName": "GS",
    "name": "Geosystems",
    "affiliationShort": "CIPR",
    "affiliationName": "Center for Integrative Petroleum Research"
  },
  {
    "groupId": "GM",
    "shortName": "GM",
    "name": "Geomechanics",
    "affiliationShort": "CIPR",
    "affiliationName": "Center for Integrative Petroleum Research"
  },
  {
    "groupId": "DA",
    "shortName": "DA",
    "name": "Data Analytics",
    "affiliationShort": "CIPR",
    "affiliationName": "Center for Integrative Petroleum Research"
  },
  {
    "groupId": "INT",
    "shortName": "INT",
    "name": "Integrative Members",
    "affiliationShort": "PETE",
    "affiliationName": "Petroleum Engineering"
  },
  {
    "groupId": "",
    "shortName": "",
    "name": "",
    "affiliationShort": "GEOL",
    "affiliationName": "Geosciences"
  },
  {
    "groupId": "",
    "shortName": "",
    "name": "",
    "affiliationShort": "MINI",
    "affiliationName": "Mining Science and Engineering"
  },
  {
    "groupId": "",
    "shortName": "",
    "name": "",
    "affiliationShort": "MATH",
    "affiliationName": "Mathematics"
  },
  {
    "groupId": "",
    "shortName": "",
    "name": "",
    "affiliationShort": "CIVIL",
    "affiliationName": "Civil and Environmental Engineering"
  },
  {
    "groupId": "",
    "shortName": "",
    "name": "",
    "affiliationShort": "CHEM",
    "affiliationName": "Chemical Engineering"
  },
  {
    "groupId": "",
    "shortName": "",
    "name": "",
    "affiliationShort": "ME",
    "affiliationName": "Mechanical Engineering"
  }
];